Refactor Torch device parameter
The old named tuple is gone and the code now lives in `cli.lab`.
